+++ client/src/components/Shared/Form/index.js
@@ -1,7 +1,6 @@
 import Form from "./Form";
 import FormInput from "./FormInput";
 import FormTextarea from "./FormTextarea";
-import FormButton from "./FormButton";
 import FormCheckbox from "./FormCheckBox";
 import FormGroup from "./FormGroup";
 import FormImageUpload from "./FormImageUpload";
@@ -10,7 +9,6 @@ export {
   Form,
   FormInput,
   FormTextarea,
-  FormButton,
   FormCheckbox,
   FormGroup,
   FormImageUpload,
